backend/profile/requirements.txt
pydantic>=2.0.0
psycopg2-binary>=2.9.0
boto3>=1.26.0
